refactored tools into a separate foldfer for clarity
